feat: implement resend verification code endpoint

+++ src/main/java/com/igniscore/api/repository/VerificationTokenRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Long> {
     Optional<VerificationToken> findByCode(String code);
     Optional<VerificationToken> findByUserId(Long userId);
-    void deleteByUserId(Long userId);
+    void deleteByUserId(Integer userId);
 }
